# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: doc-thor server API
  description: >
    Core API for doc-thor. Manages documentation projects, build jobs,
    published versions, authentication, and backend discovery.
  version: 1.0.0
servers:
  - url: /api/v1

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
# All endpoints except GET /health and POST /auth/login require a Bearer
# token.  The token is either a session token (returned by /auth/login) or
# an API key (created via POST /auth/apikey).  Both are validated the same
# way; only the raw value is ever sent over the wire — the server stores
# only its SHA-256 hash.
# ---------------------------------------------------------------------------
security:
  - bearerToken: []
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login or an API key from POST /auth/apikey.

  # -----------------------------------------------------------------
  # Reusable schemas
  # -----------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: project not found

    # --- Project ---
    Project:
      type: object
      required:
        - id
        - slug
        - name
        - source_url
        - docker_image
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: uint
        slug:
          type: string
        name:
          type: string
        source_url:
          type: string
        docker_image:
          type: string
          description: >
            Docker image the builder will run for this project.  The image
            must follow the builder contract: source repo is mounted
            read-only at /repo, generated output must be written to /output.
          example: doc-thor/builder-mkdocs
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectUpdate:
      description: >
        All fields are optional.  Only fields that are present and
        non-empty are applied.  Omit a field to leave it unchanged.
      type: object
      properties:
        name:
          type: string
        source_url:
          type: string
        docker_image:
          type: string

    # --- Build ---
    Build:
      type: object
      required:
        - id
        - project_id
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: uint
        project_id:
          type: integer
          format: uint
        ref:
          type: string
          description: Git ref that was built. Empty string when the builder uses the default branch.
        status:
          type: string
          enum: [pending, running, success, failed]
        logs:
          type: string
          description: Accumulated build log output.  Present only when non-empty.
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BuildCreate:
      type: object
      properties:
        ref:
          type: string
          description: >
            Git ref (branch, tag, or SHA) to build.  Omit or send an
            empty body to build the repository's default branch.
          example: main

    # --- Version ---
...
    VCSIntegration:
      type: object
      required:
        - id
        - name
        - provider
        - instance_url
        - enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: uint
        name:
          type: string
          description: Unique identifier for this VCS integration (e.g., "company-gitlab")
          example: company-gitlab
        provider:
          type: string
          enum: [gitlab, github, gitea]
        instance_url:
          type: string
          description: Base URL of the VCS instance
          example: https://gitlab.company.com
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    VCSIntegrationUpdate:
      type: object
      properties:
        provider:
          type: string
          enum: [gitlab, github, gitea]
        instance_url:
          type: string
        access_token:
          type: string
        webhook_secret:
          type: string
        enabled:
          type: boolean
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: database error

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /projects/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: Unique project identifier.

    get:
      summary: Get a single project
      operationId: getProject
      responses:
        "200":
          description: The requested project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update a project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete a project and all its builds and versions
      operationId: deleteProject
      responses:
        "204":
          description: Project deleted. No content returned.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # -----------------------------------------------------------------------
  # Builds
  # -----------------------------------------------------------------------
  /projects/{slug}/builds:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Trigger a build
      operationId: createBuild
      requestBody:
        required: false
        description: Body may be omitted entirely to build the default branch.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildCreate"
      responses:
        "201":
          description: Build job created with status `pending`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Project not found.
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List builds for a project
      operationId: listBuilds
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
          description: Maximum number of builds to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of builds to skip (for pagination).
      responses:
        "200":
          description: Builds ordered by created_at descending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Build"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /integrations/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get VCS integration details
      operationId: getVCSIntegration
      responses:
        "200":
          description: The requested integration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VCSIntegration"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update VCS integration
      operationId: updateVCSIntegration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VCSIntegrationUpdate"
      responses:
        "200":
          description: Updated integration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VCSIntegration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete VCS integration
      operationId: deleteVCSIntegration
      responses:
        "204":
          description: Integration deleted.
        "400":
          description: Cannot delete - projects are using this integration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /webhooks/{provider}/{slug}:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          enum: [gitlab, github, gitea]
      - name: slug
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Webhook receiver
      description: Receives webhooks from VCS platforms and triggers builds
      operationId: receiveWebhook
      security: []  # public - authenticated via webhook secret in headers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: VCS-specific webhook payload
      responses:
        "202":
          description: Webhook accepted and build triggered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [accepted, ignored]
                  build_id:
                    type: integer
                  version_tag:
                    type: string
                  ref:
                    type: string
                  auto_publish:
                    type: boolean
        "400":
          description: Project not configured for webhooks or no matching branch mapping.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid webhook signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Project not found.
          $ref: "#/components/responses/NotFound"